+++ src/deigde/deigde/shared/src/gui/native/fox/igdeNativeFoxComboBoxFilter.cpp
@@ -260,7 +260,7 @@ igdeFont *igdeNativeFoxComboBoxFilter::ComboBoxFont( const igdeComboBoxFilter &p
 	powner.GetEnvironment().GetApplicationFont( configuration );
 	
 	if( guitheme.HasProperty( igdeGuiThemePropertyNames::comboBoxFontSizeAbsolute ) ){
-		configuration.size = guitheme.GetIntProperty(
+		configuration.size = ( float )guitheme.GetIntProperty(
 			igdeGuiThemePropertyNames::comboBoxFontSizeAbsolute, 0 );
 		
 	}else if( guitheme.HasProperty( igdeGuiThemePropertyNames::comboBoxFontSize ) ){
@@ -268,7 +268,7 @@ igdeFont *igdeNativeFoxComboBoxFilter::ComboBoxFont( const igdeComboBoxFilter &p
 			igdeGuiThemePropertyNames::comboBoxFontSize, 1.0f );
 		
 	}else if( guitheme.HasProperty( igdeGuiThemePropertyNames::fontSizeAbsolute ) ){
-		configuration.size = guitheme.GetIntProperty(
+		configuration.size = ( float )guitheme.GetIntProperty(
 			igdeGuiThemePropertyNames::fontSizeAbsolute, 0 );
 		
 	}else if( guitheme.HasProperty( igdeGuiThemePropertyNames::fontSize ) ){
