+++ src/deigde/deigde/shared/src/gui/native/fox/igdeNativeFoxStatusBar.cpp
@@ -123,7 +123,7 @@ igdeFont *igdeNativeFoxStatusBar::StatusBarFont( const igdeStatusBar &powner, co
 	powner.GetEnvironment().GetApplicationFont( configuration );
 	
 	if( guitheme.HasProperty( igdeGuiThemePropertyNames::textFieldFontSizeAbsolute ) ){
-		configuration.size = guitheme.GetIntProperty(
+		configuration.size = ( float )guitheme.GetIntProperty(
 			igdeGuiThemePropertyNames::textFieldFontSizeAbsolute, 0 );
 		
 	}else if( guitheme.HasProperty( igdeGuiThemePropertyNames::textFieldFontSize ) ){
@@ -131,7 +131,7 @@ igdeFont *igdeNativeFoxStatusBar::StatusBarFont( const igdeStatusBar &powner, co
 			igdeGuiThemePropertyNames::textFieldFontSize, 1.0f );
 		
 	}else if( guitheme.HasProperty( igdeGuiThemePropertyNames::fontSizeAbsolute ) ){
-		configuration.size = guitheme.GetIntProperty(
+		configuration.size = ( float )guitheme.GetIntProperty(
 			igdeGuiThemePropertyNames::fontSizeAbsolute, 0 );
 		
 	}else if( guitheme.HasProperty( igdeGuiThemePropertyNames::fontSize ) ){
